feat: 🎸 upgrade to angular 13
✅ Closes: #24 #25 #47 #49 #58 #59 #77 #78 #83 #84
