+++ package.json
@@ -10,8 +10,9 @@
   "main": "index.js",
   "scripts": {
     "ng": "ng",
-    "ngapimock": "node test/server.js",
-    "prebuild": "rimraf dist",
+    "prengapimock": "yarn build",
+    "ngapimock": "node test/serve.js",
+    "prebuild": "rimraf dist .ngapimock",
     "build": "ng build --configuration=production",
     "lint": "ng lint",
     "start": "ng serve --proxy-config proxy.config.json --host=0.0.0.0"
@@ -32,53 +33,60 @@
   "keywords": [
     "apimock",
     "ng-apimock",
-    "test-application"
+    "test-application",
+    "http",
+    "api",
+    "scenario",
+    "mocking",
+    "endpoints",
+    "responses",
+    "delay",
+    "recording"
   ],
   "devDependencies": {
-    "@angular/animations": "12.2.3",
-    "@angular/cdk": "12.2.3",
-    "@angular/cli": "12.2.13",
-    "@angular/compiler-cli": "12.2.3",
-    "@angular/common": "12.2.3",
-    "@angular/compiler": "12.2.3",
-    "@angular/core": "12.2.3",
-    "@angular-devkit/build-angular": "12.2.3",
-    "@angular-eslint/builder": "12.3.1",
-    "@angular-eslint/eslint-plugin": "12.3.1",
-    "@angular-eslint/eslint-plugin-template": "12.3.1",
-    "@angular-eslint/schematics": "12.3.1",
-    "@angular-eslint/template-parser": "12.3.1",
-    "@angular/forms": "12.2.3",
-    "@angular/material": "12.2.3",
-    "@angular/platform-browser": "12.2.3",
-    "@angular/platform-browser-dynamic": "12.2.3",
-    "@angular/router": "12.2.3",
-    "@commitlint/cli": "13.1.0",
-    "@commitlint/config-conventional": "13.1.0",
-    "@ng-apimock/core": "3.0.3",
-    "@ng-apimock/dev-interface": "3.0.0",
-    "@semantic-release/changelog": "5.0.1",
-    "@semantic-release/commit-analyzer": "8.0.1",
-    "@semantic-release/git": "9.0.0",
-    "@semantic-release/github": "7.2.3",
-    "@semantic-release/npm": "7.1.3",
-    "@semantic-release/release-notes-generator": "9.0.3",
+    "@angular/animations": "13.0.0",
+    "@angular/cdk": "13.0.0",
+    "@angular/common": "13.0.0",
+    "@angular/compiler": "13.0.0",
+    "@angular/core": "13.0.0",
+    "@angular/forms": "13.0.0",
+    "@angular/material": "13.0.0",
+    "@angular/platform-browser": "13.0.0",
+    "@angular/platform-browser-dynamic": "13.0.0",
+    "@angular/router": "13.0.0",
+    "@angular-devkit/build-angular": "13.0.1",
+    "@angular-eslint/builder": "13.0.0-alpha.0",
+    "@angular-eslint/eslint-plugin": "13.0.0-alpha.0",
+    "@angular-eslint/eslint-plugin-template": "13.0.0-alpha.0",
+    "@angular-eslint/schematics": "13.0.0-alpha.0",
+    "@angular-eslint/template-parser": "13.0.0-alpha.0",
+    "@angular/cli": "13.0.1",
+    "@angular/compiler-cli": "13.0.0",
+    "@commitlint/cli": "14.1.0",
+    "@commitlint/config-conventional": "14.1.0",
+    "@ng-apimock/core": "3.3.0",
+    "@ng-apimock/dev-interface": "3.1.0",
+    "@semantic-release/changelog": "6.0.1",
+    "@semantic-release/commit-analyzer": "9.0.1",
+    "@semantic-release/git": "10.0.1",
+    "@semantic-release/github": "8.0.2",
+    "@semantic-release/npm": "8.0.3",
+    "@semantic-release/release-notes-generator": "10.0.2",
     "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "4.30.0",
-    "@typescript-eslint/parser": "4.30.0",
+    "@typescript-eslint/eslint-plugin": "5.3.1",
+    "@typescript-eslint/parser": "5.3.1",
     "commitizen": "4.2.4",
-    "connect": "3.7.0",
-    "eslint": "7.32.0",
+    "eslint": "8.2.0",
+    "express": "4.17.1",
     "http-proxy-middleware": "2.0.1",
-    "husky": "7.0.2",
-    "lint-staged": "11.1.2",
+    "husky": "7.0.4",
+    "lint-staged": "11.2.6",
     "rimraf": "3.0.2",
-    "rxjs": "7.3.0",
-    "semantic-release": "17.4.7",
-    "serve-static": "1.14.1",
+    "rxjs": "7.4.0",
+    "semantic-release": "18.0.0",
     "tslib": "2.3.1",
-    "ts-node": "10.2.1",
-    "typescript": "4.4.2",
+    "ts-node": "10.4.0",
+    "typescript": "4.4.4",
     "zone.js": "~0.11.4"
   },
   "lint-staged": {
