+++ src/styles.scss
@@ -1,22 +1,14 @@
-@use '~@angular/material' as mat;
+@import '~@angular/material/theming';
 
-@include mat.core();
+@include mat-core();
 
-$ngApimock-demoApplication-primary: mat.define-palette(mat.$green-palette);
-$ngApimock-demoApplication-secondary: mat.define-palette(mat.$grey-palette);
-$ngApimock-demoApplication-accent: mat.define-palette(mat.$pink-palette, A200, A100, A400);
-$ngApimock-demoApplication-warn: mat.define-palette(mat.$red-palette);
+$ngapimock-app-primary: mat-palette($mat-green);
+$ngapimock-app-secondary: mat-palette($mat-grey);
+$ngapimock-app-accent:  mat-palette($mat-pink, A200, A100, A400);
+$ngapimock-app-warn:    mat-palette($mat-red);
+$ngapimock-app-theme: mat-light-theme($ngapimock-app-primary, $ngapimock-app-accent, $ngapimock-app-warn);
 
-$ngApimock-demoApplication-theme: mat.define-light-theme((
-        color: (
-                primary: $ngApimock-demoApplication-primary,
-                accent: $ngApimock-demoApplication-accent,
-                warn: $ngApimock-demoApplication-warn,
-        )
-));
-
-@include mat.all-component-themes($ngApimock-demoApplication-theme);
-
+@include angular-material-theme($ngapimock-app-theme);
 
 html, body {
   height: 100%;
